requirements.txt
Flask
mysql-connector-python
pandas
openpyxl
scikit-learn
gunicorn
